Resolve #47: Rename artifact
