+++ gradle.properties
@@ -2,16 +2,16 @@
 org.gradle.caching=true
 org.gradle.parallel=true
 # Project properties --------------------------
+title=Scheduler.x
+description=A scheduler based on plain Vertx promise/future/periodic/timer
 version=1.0.1
 semanticVersion=-SNAPSHOT
 buildBy=local
 buildHash=
-title=Vertx Scheduler
-description=A scheduler based on plain Vertx promise/future/periodic/timer
 # SonarQube --------------------------
 systemProp.sonar.host.url=https://sonarcloud.io
 systemProp.sonar.organization=zero-88-github
-systemProp.sonar.projectKey=zero88_vertx-scheduler
+systemProp.sonar.projectKey=zero88_schedulerx
 systemProp.sonar.login=
 # Signing -----------------------------
 signing.gnupg.executable=gpg
