--- src/main/java/io/github/zero88/vertx/scheduler/impl/TaskExecutionContextInternal.java
@@ -1,13 +1,0 @@
-package io.github.zero88.vertx.scheduler.impl;
-
-import io.github.zero88.vertx.scheduler.TaskExecutionContext;
-
-public interface TaskExecutionContextInternal extends TaskExecutionContext {
-
-    void internalComplete();
-
-    Object data();
-
-    Throwable error();
-
-}
