--- src/main/java/io/github/zero88/vertx/scheduler/trigger/Trigger.java
@@ -1,15 +1,0 @@
-package io.github.zero88.vertx.scheduler.trigger;
-
-import io.github.zero88.vertx.scheduler.Task;
-
-/**
- * Represents for inspecting settings specific to a Trigger, which is used to fire a <code>{@link Task}</code> at given
- * moments in time
- *
- * @see CronTrigger
- * @see IntervalTrigger
- * @since 1.0.0
- */
-public interface Trigger {
-
-}
